fix: security hardening pass (#63)
- Extract SSRF protection to shared netutil package (SafeDialContext,
  ValidateHost, IsBlockedIP) so all HTTP clients use it
- Apply SafeDialContext to OpenAI, Claude, and webhook LLM providers
- Parameterize SQL LIMIT clauses in audit store (store.go:492, 849)
- Replace 5 error disclosure points in dashboard handlers with generic
  messages and server-side logging
- Add .dockerignore to exclude .git, keys, config, and database files
  from Docker image layers
