+++ .dockerignore
@@ -1,0 +1,24 @@
+.git
+.github
+*.key
+keys/
+oktsec.yaml
+oktsec.yaml.bak
+*.db
+*.db-shm
+*.db-wal
+.env*
+DOCS/
+CLAUDE.md
+.claude/
+demo/
+testdata/
+coverage.out
+dist/
+node_modules/
+sdk/python/.venv/
+sdk/python/*.egg-info/
+sdk/python/dist/
+__pycache__/
+site/
+.DS_Store
